(tae): small code refactor 2
