+++ pkg/vm/engine/tae/tables/updates/delete.go
